+++ CHANGES.rst
@@ -7,6 +7,7 @@ Unreleased Changes
 * Fix ``/pay_period_for`` redirect to be a 302 instead of 301, add redirect logging, remove some old debug logging from that view.
 * Fix logging exception in db_event_handlers on initial data load.
 * Switch ofxparse requirement to use upstream repo now that https://github.com/jseutter/ofxparse/pull/127 is merged.
+* `Issue #83 <https://github.com/jantman/biweeklybudget/issues/83>`_ - Fix 500 error preventing display of balance chart on ``/`` view when an account has a None ledger balance.
 
 0.1.2 (2017-05-28)
 ------------------